Add files via upload
HTML-CSS Navbar Project: Explore the basics of web development with this repository showcasing a straightforward website. Dive into the essentials of HTML and CSS as you construct a user-friendly webpage, complete with a responsive navigation bar for seamless interaction. Perfect for beginners seeking hands-on experience in building and styling a foundational web project.